move trade page to separate path and adapt navbar
